// src/api/schemas/authorSchema.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keyo API
 * API for Keyo Polls app
 * OpenAPI spec version: 1.0.0
 */
import type { AuthorSchemaAvatar } from './authorSchemaAvatar';

export interface AuthorSchema {
  id: number;
  username: string;
  display_name: string;
  total_aura: number;
  avatar?: AuthorSchemaAvatar;
}